requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-six==1.0.4
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.3
psycopg2==2.8.6
pylint==2.6.0
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
xlwt==1.3.0
